+++ crates/services/importer/src/ports.rs
@@ -41,13 +41,10 @@ use fuel_core_types::{
 
 #[cfg_attr(test, mockall::automock(type Database = crate::importer::test::MockDatabase;))]
 /// The executors port.
-pub trait Executor: Send + Sync {
+pub trait Validator: Send + Sync {
     /// Executes the block and returns the result of execution with uncommitted database
     /// transaction.
-    fn execute_without_commit(
-        &self,
-        block: Block,
-    ) -> ExecutorResult<UncommittedResult<Changes>>;
+    fn validate(&self, block: Block) -> ExecutorResult<UncommittedResult<Changes>>;
 }
 
 /// The trait indicates that the type supports storage transactions.